Move tests into src folder. Can now build the project with npm run build
