+++ package.json
@@ -78,11 +78,11 @@
   "repository": "dcathcart/dx-cop",
   "scripts": {
     "build": "tsc -p .",
-    "lint": "eslint src/**/*.ts test/**/*.ts",
+    "lint": "eslint src/**/*.ts",
     "postpack": "rimraf oclif.manifest.json",
-    "posttest": "eslint src/**/*.ts test/**/*.ts",
+    "posttest": "eslint src/**/*.ts",
     "prepack": "rimraf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
-    "test": "mocha -r ts-node/register \"test/**/*.test.ts\"",
+    "test": "mocha -r ts-node/register \"src/test/**/*.test.ts\"",
     "test-old": "nyc --extension .ts --require ts-node/register mocha --forbid-only \"test/**/*.test.ts\"",
     "version": "oclif-dev readme && git add README.md"
   }
