+++ src/test/check/EmailToCaseSettingsChecker.test.ts
@@ -1,14 +1,14 @@
 import 'mocha';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
-import { EmailToCaseSettings } from '../../src/metadata_browser/EmailToCaseSettings';
-import { EmailToCaseSettingsChecker } from '../../src/check/EmailToCaseSettingsChecker';
-import { SfdxProjectBrowser } from '../../src/metadata_browser/SfdxProjectBrowser';
+import { EmailToCaseSettings } from '../../metadata_browser/EmailToCaseSettings';
+import { EmailToCaseSettingsChecker } from '../../check/EmailToCaseSettingsChecker';
+import { SfdxProjectBrowser } from '../../metadata_browser/SfdxProjectBrowser';
 
 describe('EmailToCaseSettingsChecker', () => {
   describe('.run() method', () => {
     it('should return an array of metadata problems', () => {
-      const emailToCaseSettings = new EmailToCaseSettings('test/fixtures/settings/Case-errors.settings-meta.xml');
+      const emailToCaseSettings = new EmailToCaseSettings('src/test/fixtures/settings/Case-errors.settings-meta.xml');
       const sfdxProjectBrowser = new SfdxProjectBrowser(null);
       const mockProjectBrowser = sinon.mock(sfdxProjectBrowser);
       mockProjectBrowser.expects('emailToCaseSettings').once().returns(emailToCaseSettings);
@@ -19,7 +19,7 @@ describe('EmailToCaseSettingsChecker', () => {
       mockProjectBrowser.verify();
     });
     it('should return an empty array when there are no problems', () => {
-      const emailToCaseSettings = new EmailToCaseSettings('test/fixtures/settings/Case.settings-meta.xml');
+      const emailToCaseSettings = new EmailToCaseSettings('src/test/fixtures/settings/Case.settings-meta.xml');
       const sfdxProjectBrowser = new SfdxProjectBrowser(null);
       const mockProjectBrowser = sinon.mock(sfdxProjectBrowser);
       mockProjectBrowser.expects('emailToCaseSettings').once().returns(emailToCaseSettings);
@@ -33,7 +33,7 @@ describe('EmailToCaseSettingsChecker', () => {
 
   describe('.checkEmailServicesAddress() method', () => {
     it('should return a metadata error when an <emailServicesAddress> field exists', () => {
-      const fileName = 'test/fixtures/settings/Case-errors.settings-meta.xml';
+      const fileName = 'src/test/fixtures/settings/Case-errors.settings-meta.xml';
       const emailToCaseSettings = new EmailToCaseSettings(fileName);
       const checker = new EmailToCaseSettingsChecker(null);
       const result = checker['checkEmailServicesAddress'](emailToCaseSettings);
@@ -46,7 +46,7 @@ describe('EmailToCaseSettingsChecker', () => {
     });
 
     it('should return an empty array when there are no errors', () => {
-      const emailToCaseSettings = new EmailToCaseSettings('test/fixtures/settings/Case.settings-meta.xml');
+      const emailToCaseSettings = new EmailToCaseSettings('src/test/fixtures/settings/Case.settings-meta.xml');
       const checker = new EmailToCaseSettingsChecker(null);
       const result = checker['checkEmailServicesAddress'](emailToCaseSettings);
       expect(result.length).to.equal(0);
@@ -55,7 +55,7 @@ describe('EmailToCaseSettingsChecker', () => {
 
   describe('.checkIsVerified() method', () => {
     it('should return a metadata error when an <isVerified> field exists', () => {
-      const fileName = 'test/fixtures/settings/Case-errors.settings-meta.xml';
+      const fileName = 'src/test/fixtures/settings/Case-errors.settings-meta.xml';
       const emailToCaseSettings = new EmailToCaseSettings(fileName);
       const checker = new EmailToCaseSettingsChecker(null);
       const result = checker['checkIsVerified'](emailToCaseSettings);
@@ -68,7 +68,7 @@ describe('EmailToCaseSettingsChecker', () => {
     });
 
     it('should return an empty array when there are no errors', () => {
-      const emailToCaseSettings = new EmailToCaseSettings('test/fixtures/settings/Case.settings-meta.xml');
+      const emailToCaseSettings = new EmailToCaseSettings('src/test/fixtures/settings/Case.settings-meta.xml');
       const checker = new EmailToCaseSettingsChecker(null);
       const result = checker['checkIsVerified'](emailToCaseSettings);
       expect(result.length).to.equal(0);
@@ -77,7 +77,7 @@ describe('EmailToCaseSettingsChecker', () => {
 
   describe('.checkSortOrder() method', () => {
     it("should return a metadata error when <routingAddresses> aren't sorted by <routingName>", () => {
-      const fileName = 'test/fixtures/settings/Case-errors.settings-meta.xml';
+      const fileName = 'src/test/fixtures/settings/Case-errors.settings-meta.xml';
       const emailToCaseSettings = new EmailToCaseSettings(fileName);
       const checker = new EmailToCaseSettingsChecker(null);
       const result = checker['checkSortOrder'](emailToCaseSettings);
@@ -92,7 +92,7 @@ describe('EmailToCaseSettingsChecker', () => {
     });
 
     it('should return an empty array when there are no errors', () => {
-      const emailToCaseSettings = new EmailToCaseSettings('test/fixtures/settings/Case.settings-meta.xml');
+      const emailToCaseSettings = new EmailToCaseSettings('src/test/fixtures/settings/Case.settings-meta.xml');
       const checker = new EmailToCaseSettingsChecker(null);
       const result = checker['checkSortOrder'](emailToCaseSettings);
       expect(result.length).to.equal(0);
