+++ src/test/check/RecordTypePicklistChecker.test.ts
@@ -1,10 +1,10 @@
 import 'mocha';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
-import { PicklistField } from '../../src/metadata_browser/PicklistField';
-import { RecordType } from '../../src/metadata_browser/RecordType';
-import { RecordTypePicklistChecker } from '../../src/check/RecordTypePicklistChecker';
-import { SfdxProjectBrowser } from '../../src/metadata_browser/SfdxProjectBrowser';
+import { PicklistField } from '../../metadata_browser/PicklistField';
+import { RecordType } from '../../metadata_browser/RecordType';
+import { RecordTypePicklistChecker } from '../../check/RecordTypePicklistChecker';
+import { SfdxProjectBrowser } from '../../metadata_browser/SfdxProjectBrowser';
 
 describe('RecordTypePicklistChecker', () => {
   // this one is less unit test, more functional test
@@ -13,13 +13,13 @@ describe('RecordTypePicklistChecker', () => {
     it('should return an array of metadata problems', () => {
       const objectNames = ['TestObject1'];
       const picklistFields = [
-        new PicklistField('test/fixtures/objects/TestObject1/fields/CustomPicklist1__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject1/fields/CustomPicklist2__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject1/fields/StandardPicklist1.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject1/fields/StandardPicklist2.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/CustomPicklist1__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/CustomPicklist2__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/StandardPicklist1.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/StandardPicklist2.field-meta.xml'),
       ];
       const recordTypes = [
-        new RecordType('test/fixtures/objects/TestObject1/recordTypes/BadExample.recordType-meta.xml'),
+        new RecordType('src/test/fixtures/objects/TestObject1/recordTypes/BadExample.recordType-meta.xml'),
       ];
       const sfdxProjectBrowser = new SfdxProjectBrowser(null);
       const mockProjectBrowser = sinon.mock(sfdxProjectBrowser);
@@ -36,13 +36,13 @@ describe('RecordTypePicklistChecker', () => {
     it('should return an empty array when there are no problems', () => {
       const objectNames = ['TestObject1'];
       const picklistFields = [
-        new PicklistField('test/fixtures/objects/TestObject1/fields/CustomPicklist1__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject1/fields/CustomPicklist2__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject1/fields/StandardPicklist1.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject1/fields/StandardPicklist2.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/CustomPicklist1__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/CustomPicklist2__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/StandardPicklist1.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject1/fields/StandardPicklist2.field-meta.xml'),
       ];
       const recordTypes = [
-        new RecordType('test/fixtures/objects/TestObject1/recordTypes/GoodExample.recordType-meta.xml'),
+        new RecordType('src/test/fixtures/objects/TestObject1/recordTypes/GoodExample.recordType-meta.xml'),
       ];
       const sfdxProjectBrowser = new SfdxProjectBrowser(null);
       const mockProjectBrowser = sinon.mock(sfdxProjectBrowser);
