+++ src/test/check/RecordTypePicklistValueChecker.test.ts
@@ -1,10 +1,10 @@
 import 'mocha';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
-import { PicklistField } from '../../src/metadata_browser/PicklistField';
-import { RecordType } from '../../src/metadata_browser/RecordType';
-import { RecordTypePicklistValueChecker } from '../../src/check/RecordTypePicklistValueChecker';
-import { SfdxProjectBrowser } from '../../src/metadata_browser/SfdxProjectBrowser';
+import { PicklistField } from '../../metadata_browser/PicklistField';
+import { RecordType } from '../../metadata_browser/RecordType';
+import { RecordTypePicklistValueChecker } from '../../check/RecordTypePicklistValueChecker';
+import { SfdxProjectBrowser } from '../../metadata_browser/SfdxProjectBrowser';
 
 describe('RecordTypePicklistValueChecker', () => {
   // this one is less unit test, more functional test
@@ -13,13 +13,13 @@ describe('RecordTypePicklistValueChecker', () => {
     it('should return an array of metadata problems', () => {
       const objectNames = ['TestObject2'];
       const picklistFields = [
-        new PicklistField('test/fixtures/objects/TestObject2/fields/CustomPicklist1__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject2/fields/CustomPicklist2__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject2/fields/GloablPicklist1.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject2/fields/StandardPicklist1.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/CustomPicklist1__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/CustomPicklist2__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/GloablPicklist1.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/StandardPicklist1.field-meta.xml'),
       ];
       const recordTypes = [
-        new RecordType('test/fixtures/objects/TestObject2/recordTypes/BadExample.recordType-meta.xml'),
+        new RecordType('src/test/fixtures/objects/TestObject2/recordTypes/BadExample.recordType-meta.xml'),
       ];
       const sfdxProjectBrowser = new SfdxProjectBrowser(null);
       const mockProjectBrowser = sinon.mock(sfdxProjectBrowser);
@@ -36,13 +36,13 @@ describe('RecordTypePicklistValueChecker', () => {
     it('should return an empty array when there are no problems', () => {
       const objectNames = ['TestObject2'];
       const picklistFields = [
-        new PicklistField('test/fixtures/objects/TestObject2/fields/CustomPicklist1__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject2/fields/CustomPicklist2__c.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject2/fields/GloablPicklist1.field-meta.xml'),
-        new PicklistField('test/fixtures/objects/TestObject2/fields/StandardPicklist1.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/CustomPicklist1__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/CustomPicklist2__c.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/GloablPicklist1.field-meta.xml'),
+        new PicklistField('src/test/fixtures/objects/TestObject2/fields/StandardPicklist1.field-meta.xml'),
       ];
       const recordTypes = [
-        new RecordType('test/fixtures/objects/TestObject2/recordTypes/GoodExample.recordType-meta.xml'),
+        new RecordType('src/test/fixtures/objects/TestObject2/recordTypes/GoodExample.recordType-meta.xml'),
       ];
       const sfdxProjectBrowser = new SfdxProjectBrowser(null);
       const mockProjectBrowser = sinon.mock(sfdxProjectBrowser);
