+++ src/test/fixtures/lwc/BadExample/BadExample.js-meta.xml
