+++ src/test/fixtures/objects/TestObject1/fields/CustomPicklist2__c.field-meta.xml
