+++ src/test/fixtures/objects/TestObject1/fields/StandardPicklist1.field-meta.xml
