+++ src/test/fixtures/objects/TestObject1/fields/StandardPicklist2.field-meta.xml
