+++ src/test/fixtures/objects/TestObject1/recordTypes/BadExample.recordType-meta.xml
