+++ src/test/fixtures/objects/TestObject1/recordTypes/GoodExample.recordType-meta.xml
