+++ src/test/fixtures/objects/TestObject2/fields/GlobalPicklist1.field-meta.xml
