+++ src/test/fixtures/objects/TestObject2/fields/StandardPicklist1.field-meta.xml
