+++ src/test/fixtures/objects/TestObject2/recordTypes/BadExample.recordType-meta.xml
