+++ src/test/fixtures/objects/TestObject2/recordTypes/GoodExample.recordType-meta.xml
