+++ tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "lib",
     "rootDir": "src"
   },
-  "include": ["./src/**/*.ts", "./test/**/*.test.ts"]
+  "include": ["./src/**/*.ts"]
 }
